Fixed a bug in src/transitions.c:cell_grid_animation()
- the bug was in src/tranisitions.c:cell_grid_animation() and it was at the action of "the 2 right
internal border-columns of the grid get drawn from bottom to top" the row we were starting printing
from was ((bottom+2) - i) with i going from top to bottom, but this made us always start printing at
a constant position (shifted upwards from the bottom by a certain amount depending on the size of the
window, I did not see the bug because my window size was perfectly sized for the bug to trick me,
heheheha) (the constant row position was around 55 the MENU_HEIGHT), because
    bottom - top = menu_start_row + 1 - (menu_start_row + MENU_HEIGHT-1) = MENU_HEIGHT plus or minus smth
                    the above values are from call of the animation function
        notice that the variable menu_start_row that centers the stuff gets canceled out, which is
        what caused the bug
plusORminus smth doesnt matter, what matters is that we always start printing from a constant position
which is not what we should have for centering. What we had to do is start from bottom and keep subtracting
a counter that starts from 0 and goes all the way up to MENU_HEIGHT. and this solved the the bug.
debugged by Oabdullae and Haytam.                                                                `#`